it was moved auxiliary classes - TranslateLanguage and TranslateresultImpl - into dto-package
