+++ src/main/java/evg/testt/dto/TranslateResultImpl.java
@@ -1,4 +1,4 @@
-package evg.testt.service.impl.translateserviceimpl;
+package evg.testt.dto;
 
 import evg.testt.service.translateservice.TranslateResult;
