+++ src/main/resources/database_rdbms.dev.properties
@@ -4,7 +4,7 @@ jdbc.platform=MYSQL
 jdbc.driverClassName=com.mysql.jdbc.Driver
 jdbc.databaseUrl=jdbc:mysql://localhost:3306/testT?characterEncoding=UTF-8
 jdbc.username=root
-jdbc.password=1
+jdbc.password=root
 
 connection.pool_size=10
 hibernate.hbm2ddl.auto=create
